+++ source/atelier/ui/list/gridlist.d
@@ -153,7 +153,7 @@ class GridList: GuiElement {
 		_container.container.removeChildrenGuis();
 	}
 
-	override void removeChildGui(uint id) {
+	override void removeChildGui(size_t id) {
 		_container.container.removeChildGui(id);
 		_nbElements = _container.container.getChildrenGuisCount();
 		_idElementSelected = 0u;
